+++ lendo-backend-searchIn-googlePlace/lendo-backend-searchIn-googlePlace-component/lendo-backend-searchIn-googlePlace-component-api/src/main/java/se/lendo/test/search/backend/component/clinet/SearchPlaceClientImp.java
@@ -22,7 +22,7 @@ public class SearchPlaceClientImp implements SearchPlaceClient {
     }
 
     @Override
-    public List<Place> getPlacesByQueryBasedOnRadar(String query, int limit, double radius, Param... extraParams) {
+    public List<Place> getPlacesByQueryBasedOnRadar(String query, int limit, int radius, Param... extraParams) {
         return searchPlaceClient.getPlacesByQueryBasedOnRadar(query, limit, radius, extraParams);
     }
 }
