+++ lendo-backend-searchIn-googlePlace/lendo-backend-searchIn-googlePlace-component/lendo-backend-searchIn-googlePlace-component-service-impl/src/main/java/se/lendo/test/backend/googlePlaceApi/component/service/SearchPlaceServiceImpl.java
@@ -24,7 +24,7 @@ public class SearchPlaceServiceImpl implements SearchPlaceService {
     }
 
     @Override
-    public List<Place> getPlacesByQueryBasedOnRadar(String query, int limit, double radius, Param... extraParams) {
+    public List<Place> getPlacesByQueryBasedOnRadar(String query, int limit, int radius, Param... extraParams) {
 
         return placeApiSearchClient.getPlacesByQueryBasedOnRadar(query, limit, radius, extraParams);
