+++ lendo-integration-googleplace-api/lendo-integration-googleplace-api-client/src/main/java/se/lendo/test/intergration/googleplace/api/client/GooglePlaceApiSearchClient.java
@@ -365,7 +365,7 @@ public interface GooglePlaceApiSearchClient {
 
     List<Place> getPlacesByQuery(String query, int limit, Param... extraParams);
 
-    List<Place> getPlacesByQueryBasedOnRadar(String query, int limit, double radius,Param... extraParams);
+    List<Place> getPlacesByQueryBasedOnRadar(String query, int limit, int radius,Param... extraParams);
 
     // create a class with builder class that have lat and all other
     List<Place> getNearbyPlaces(double lat, double lng, double radius, int limit, Param... extraParams);
